CCSS-18: Move Implementations to the Library Directory
