+++ Src/Client/Main.cpp
@@ -1,8 +1,8 @@
 #include <cstring>
 #include <iostream>
-
-#include "Client.hpp"
 #include <unistd.h>
+
+#include "../Library/Client.hpp"
 
 int main()
 {
